Move EditorStyle into editor module
